+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Added support for resolving superclass properties for not-NSObject subclasses
 
+### Bug Fixes
+
+- Fixed rendering `{{ block.super }}` with several levels of inheritance
+
+
 ## 0.10.1
 
 ### Enhancements