+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Enhancements
 
 - Added support for resolving superclass properties for not-NSObject subclasses
+- Various improvements for for tag:
+  - added support for iterating over tuple elements and subscripting it with label or index
+  - added support for iterating array with index and value
+  - fixed using spaces between loop variables
+  - added support for subscripting dictionary using loop variable for key
+
 
 ### Bug Fixes